Parse XML and display on the dashboard
