--- application/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "processa-processo",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "camaro": "^6.1.0",
-    "express": "^4.17.1",
-    "fast-xml-parser": "^3.19.0",
-    "he": "^1.2.0"
-  }
-}
