+++ client/src/components/Row.js
@@ -1,0 +1,73 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { makeStyles } from '@material-ui/core/styles';
+import Box from '@material-ui/core/Box';
+import Collapse from '@material-ui/core/Collapse';
+import IconButton from '@material-ui/core/IconButton';
+import Table from '@material-ui/core/Table';
+import TableBody from '@material-ui/core/TableBody';
+import TableCell from '@material-ui/core/TableCell';
+import TableHead from '@material-ui/core/TableHead';
+import TableRow from '@material-ui/core/TableRow';
+import Typography from '@material-ui/core/Typography';
+import KeyboardArrowDownIcon from '@material-ui/icons/KeyboardArrowDown';
+import KeyboardArrowUpIcon from '@material-ui/icons/KeyboardArrowUp';
+import RowChild from './RowChild';
+
+const useRowStyles = makeStyles({
+    root: {
+      '& > *': {
+        borderBottom: 'unset',
+      },
+    },
+  });
+
+const propTypes = {
+    row: PropTypes.shape({
+        claimed: PropTypes.string.isRequired,
+        court: PropTypes.string.isRequired,
+        laywer: PropTypes.string.isRequired,
+        history: PropTypes.arrayOf(
+        PropTypes.shape({
+            amount: PropTypes.number.isRequired,
+            customerId: PropTypes.string.isRequired,
+            date: PropTypes.string.isRequired,
+        }),
+        ).isRequired,
+        name: PropTypes.string.isRequired,
+        value: PropTypes.number.isRequired,
+        processNumber: PropTypes.number.isRequired,
+    }).isRequired,
+};
+
+export default function Row(props) {
+    const { row } = props;
+    const [open, setOpen] = React.useState(false);
+    const classes = useRowStyles();
+
+    return (
+      <React.Fragment>
+        <TableRow className={classes.root}>
+          <TableCell>
+            <IconButton aria-label="expand row" size="small" onClick={() => setOpen(!open)}>
+              {open ? <KeyboardArrowUpIcon /> : <KeyboardArrowDownIcon />}
+            </IconButton>
+          </TableCell>
+          <TableCell component="th" scope="row">
+            {row.name}
+          </TableCell>
+          <TableCell>{row.claimed}</TableCell>
+          <TableCell>{row.advogado}</TableCell>
+          <TableCell>{row.court}</TableCell>
+          <TableCell>{row.processNumber}</TableCell>
+        </TableRow>
+        <TableRow>
+          <TableCell style={{ paddingBottom: 0, paddingTop: 0 }} colSpan={6}>
+            <Collapse in={open} timeout="auto" unmountOnExit>
+                <RowChild row={row} />
+            </Collapse>
+          </TableCell>
+        </TableRow>
+      </React.Fragment>
+    );
+  }
