--- client/src/routes/Dashboard.js
@@ -1,156 +1,0 @@
-import React from "react";
-import useTitle from "../hooks/use-title";
-import PropTypes from 'prop-types';
-import { makeStyles } from '@material-ui/core/styles';
-import Box from '@material-ui/core/Box';
-import Collapse from '@material-ui/core/Collapse';
-import IconButton from '@material-ui/core/IconButton';
-import Table from '@material-ui/core/Table';
-import TableBody from '@material-ui/core/TableBody';
-import TableCell from '@material-ui/core/TableCell';
-import TableContainer from '@material-ui/core/TableContainer';
-import TableHead from '@material-ui/core/TableHead';
-import TableRow from '@material-ui/core/TableRow';
-import Typography from '@material-ui/core/Typography';
-import Paper from '@material-ui/core/Paper';
-import KeyboardArrowDownIcon from '@material-ui/icons/KeyboardArrowDown';
-import KeyboardArrowUpIcon from '@material-ui/icons/KeyboardArrowUp';
-
-const useRowStyles = makeStyles({
-    root: {
-      '& > *': {
-        borderBottom: 'unset',
-      },
-    },
-  });
-  
-  function createData(name, reclamadas, advogado, vara, processo, price) {
-    return {
-      name,
-      reclamadas,
-      advogado,
-      vara,
-      processo,
-      price,
-      history: [
-        { date: '2020-01-05', customerId: '11091700', amount: 3 },
-        { date: '2020-01-02', customerId: 'Anonymous', amount: 1 },
-      ],
-    };
-  }
-  
-  function Row(props) {
-    const { row } = props;
-    const [open, setOpen] = React.useState(false);
-    const classes = useRowStyles();
-  
-    return (
-      <React.Fragment>
-        <TableRow className={classes.root}>
-          <TableCell>
-            <IconButton aria-label="expand row" size="small" onClick={() => setOpen(!open)}>
-              {open ? <KeyboardArrowUpIcon /> : <KeyboardArrowDownIcon />}
-            </IconButton>
-          </TableCell>
-          <TableCell component="th" scope="row">
-            {row.name}
-          </TableCell>
-          <TableCell>{row.reclamadas}</TableCell>
-          <TableCell>{row.advogado}</TableCell>
-          <TableCell>{row.vara}</TableCell>
-          <TableCell>{row.processo}</TableCell>
-        </TableRow>
-        <TableRow>
-          <TableCell style={{ paddingBottom: 0, paddingTop: 0 }} colSpan={6}>
-            <Collapse in={open} timeout="auto" unmountOnExit>
-              <Box margin={1}>
-                <Typography variant="h6" gutterBottom component="div">
-                  Detalhes
-                </Typography>
-                <Table size="small" aria-label="purchases">
-                  <TableHead>
-                    <TableRow>
-                      <TableCell>Date</TableCell>
-                      <TableCell>Customer</TableCell>
-                      <TableCell align="right">Amount</TableCell>
-                      <TableCell align="right">Total price ($)</TableCell>
-                    </TableRow>
-                  </TableHead>
-                  <TableBody>
-                    {row.history.map((historyRow) => (
-                      <TableRow key={historyRow.date}>
-                        <TableCell component="th" scope="row">
-                          {historyRow.date}
-                        </TableCell>
-                        <TableCell>{historyRow.customerId}</TableCell>
-                        <TableCell align="right">{historyRow.amount}</TableCell>
-                        <TableCell align="right">
-                          {Math.round(historyRow.amount * row.price * 100) / 100}
-                        </TableCell>
-                      </TableRow>
-                    ))}
-                  </TableBody>
-                </Table>
-              </Box>
-            </Collapse>
-          </TableCell>
-        </TableRow>
-      </React.Fragment>
-    );
-  }
-  
-  Row.propTypes = {
-    row: PropTypes.shape({
-      reclamadas: PropTypes.string.isRequired,
-      vara: PropTypes.string.isRequired,
-      advogado: PropTypes.string.isRequired,
-      history: PropTypes.arrayOf(
-        PropTypes.shape({
-          amount: PropTypes.number.isRequired,
-          customerId: PropTypes.string.isRequired,
-          date: PropTypes.string.isRequired,
-        }),
-      ).isRequired,
-      name: PropTypes.string.isRequired,
-      price: PropTypes.number.isRequired,
-      processo: PropTypes.number.isRequired,
-    }).isRequired,
-  };
-  
-  const rows = [
-    createData('ACEP', 'BK Consultoria e Serviços Ltda.', 'PRC', '11\u00AA', '0101871-50.2017.5.01.0026', 3.99),
-    createData('ACEP', 'BK Consultoria e Serviços Ltda.', 'PRC', '11\u00AA', '0101871-50.2017.5.01.0026', 3.99),
-    createData('ACEP', 'BK Consultoria e Serviços Ltda.', 'PRC', '11\u00AA', '0101871-50.2017.5.01.0026', 3.99),
-    createData('ACEP', 'BK Consultoria e Serviços Ltda.', 'PRC', '11\u00AA', '0101871-50.2017.5.01.0026', 3.99),
-    createData('ACEP', 'BK Consultoria e Serviços Ltda.', 'PRC', '11\u00AA', '0101871-50.2017.5.01.0026', 3.99),
-  ];
-  
-
-export default function Dashboard() {
-    useTitle("Dashboard");
-
-    return (
-        <div>
-            <h1>Dashboard</h1>
-            <TableContainer component={Paper}>
-                <Table aria-label="collapsible table">
-                    <TableHead>
-                    <TableRow>
-                        <TableCell />
-                        <TableCell><b>Reclamante/Requerente</b></TableCell>
-                        <TableCell><b>Reclamadas</b></TableCell>
-                        <TableCell><b>ADV</b></TableCell>
-                        <TableCell><b>Vara</b></TableCell>
-                        <TableCell><b>Processo</b></TableCell>
-                    </TableRow>
-                    </TableHead>
-                    <TableBody>
-                    {rows.map((row) => (
-                        <Row key={row.name} row={row} />
-                    ))}
-                    </TableBody>
-                </Table>
-            </TableContainer>
-        </div>
-    );
-}
